feat(settings): add StorageSettingsScreen and integrate into navigation
- Implemented StorageSettingsScreen with cache management options.
- Updated UserManagementScreen to include background color for consistency.
- Enhanced SettingsNavigation to include routes for new settings screens: Appearance, Audio, Backup, Language, Notification, Security, and Storage settings.
