+++ core/ui/src/main/java/com/raulshma/jellyplay/core/ui/navigation/NavKey.kt
@@ -51,6 +51,14 @@ sealed class Route : NavKey {
     @Serializable data object ServerManagement : Route()
     @Serializable data object UserManagement : Route()
     @Serializable data object SeerrSettings : Route()
+    @Serializable data object AppearanceSettings : Route()
+    @Serializable data object PlaybackSettings : Route()
+    @Serializable data object AudioSettings : Route()
+    @Serializable data object LanguageSettings : Route()
+    @Serializable data object NotificationSettings : Route()
+    @Serializable data object StorageSettings : Route()
+    @Serializable data object SecuritySettings : Route()
+    @Serializable data object BackupSettings : Route()
 
     @Serializable data object Artists : Route()
     @Serializable data object Albums : Route()
